path trouble

+++ internal/model/forum.go
@@ -1,8 +1,8 @@
 package model
 
 type Forum struct {
-	ID      int64  `json:"-"`
-	Posts   int64  `json:"posts"`
+	ID      int32  `json:"-"`
+	Posts   int32  `json:"posts"`
 	Slug    string `json:"slug"`
 	Threads int32  `json:"threads"`
 	Title   string `json:"title"`
