+++ internal/model/post.go
@@ -3,17 +3,17 @@ package model
 import "time"
 
 type Post struct {
-	ID            int64   `json:"id,omitempty"`
-	Parent        int64   `json:"parent"`
-	Thread        int32   `json:"thread,omitempty"`
-	Forum         string  `json:"forum,omitempty"`
-	Author        string  `json:"author"`
-	Created       time.Time  `json:"created,omitempty"`
-	IsEdited      bool    `json:"isEdited"`
-	Message       string  `json:"message"`
-	Path          []int64 `json:"-"`
-	Childs        []Post   `json:"childs,omitempty"`
-	ParentPointer *Post   `json:"-"`
+	ID            int32     `json:"id,omitempty"`
+	Parent        int32     `json:"parent"`
+	Thread        int32     `json:"thread,omitempty"`
+	Forum         string    `json:"forum,omitempty"`
+	Author        string    `json:"author"`
+	Created       time.Time `json:"created,omitempty"`
+	IsEdited      bool      `json:"isEdited"`
+	Message       string    `json:"message"`
+	Path          []int32   `json:"-"`
+	Childs        []Post    `json:"childs,omitempty"`
+	ParentPointer *Post     `json:"-"`
 }
 
 type PostFull struct {
@@ -22,7 +22,6 @@ type PostFull struct {
 	Post   *Post   `json:"post,omitempty"`
 	Thread *Thread `json:"thread,omitempty"`
 }
-
 
 type PostUpdate struct {
 	Message string `json:"message,omitempty"`
